Improved debug fallback for TypeScript parser.
