+++ package.json
@@ -31,7 +31,7 @@
     "devDependencies": {
         "@types/mkdirp": "^0.5.2",
         "@types/node": "^12.12.3",
-        "typescript": "~3.7.4"
+        "typescript": "~3.7.5"
     },
     "files": [
         "api-docs",
