+++ typescript/sources/Project.ts
@@ -27,6 +27,43 @@ export class Project {
         return new M.FileMember(child);
     }
 
+    public static loadFromArguments(args: Array<string>): Project {
+
+        const parsedCommandLine = TS.parseCommandLine(args);
+
+        return new Project(
+            TS.createProgram(
+                parsedCommandLine.fileNames,
+                parsedCommandLine.options
+            )
+        );
+    }
+
+    public static loadFromDirectory(directoryPath: string): Project {
+
+        const tsConfig = TS.readJsonConfigFile(
+            TS.sys.resolvePath(directoryPath),
+            TS.sys.readFile
+        );
+
+        const parsedCommandLine = TS.parseJsonConfigFileContent(
+            tsConfig,
+            TS.sys,
+            directoryPath
+        );
+
+        const project = new Project(
+            TS.createProgram(
+                parsedCommandLine.fileNames,
+                parsedCommandLine.options
+            )
+        );
+
+        project.directoryPath = directoryPath;
+
+        return project;
+    }
+
     /* *
      *
      *  Constructor
